requirements.txt
boto3>=1.25.0
python-dateutil>=2.8.0
freezegun>=1.2.0  # Required for testing
pytest>=7.0.0     # For running tests
pytest-cov>=4.0.0 # For test coverage reporting